el usuario al crear evento aparece correctamente
